backend/requirements.txt
fastapi==0.124.2
google-genai==1.57.0
uvicorn[standard]==0.38.0
pydantic==2.12.5
httpx==0.28.1
python-dotenv==1.2.2
pytest==9.0.2
